requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==3.2.9
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-mptt==0.13.4
django-quill-editor==0.1.40
django-taggit==2.1.0
django-widget-tweaks==1.4.12
fontawesomefree==6.0.0
importlib-metadata==4.11.1
packaging==21.3
Pillow==9.0.1
Pint==0.18
pyparsing==3.0.7
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
zipp==3.7.0
